# optimize: TEA encryption