--- a/binary/tea.go
+++ b/binary/tea.go
@@ -10,6 +10,7 @@ type TEA [4]uint32
 // Encrypt tea 加密
 // http://bbs.chinaunix.net/thread-583468-1-1.html
 // 感谢xichen大佬对TEA的解释
+//go:nosplit
 func (t TEA) Encrypt(src []byte) (dst []byte) {
 	lens := len(src)
 	fill := 10 - (lens+1)%8
@@ -19,69 +20,78 @@ func (t TEA) Encrypt(src []byte) (dst []byte) {
 	copy(dst[fill:], src)
 
 	var iv1, iv2, holder uint64
+	t0, t1, t2, t3 := t[0], t[1], t[2], t[3]
 	for i := 0; i < len(dst); i += 8 {
-		block := binary.BigEndian.Uint64(dst[i:])
-		holder = block ^ iv1
-		iv1 = t.encode(holder)
-		iv1 = iv1 ^ iv2
+		holder = binary.BigEndian.Uint64(dst[i:]) ^ iv1
+		iv1 = enctea(t0, t1, t2, t3, holder) ^ iv2
 		iv2 = holder
 		binary.BigEndian.PutUint64(dst[i:], iv1)
 	}
-
 	return dst
 }
 
+//go:nosplit
 func (t TEA) Decrypt(data []byte) []byte {
 	if len(data) < 16 || len(data)%8 != 0 {
 		return nil
 	}
 	dst := make([]byte, len(data))
-	var iv1, iv2, holder, tmp uint64
+
+	var iv1, iv2, holder uint64
+	t0, t1, t2, t3 := t[0], t[1], t[2], t[3]
 	for i := 0; i < len(dst); i += 8 {
-		block := binary.BigEndian.Uint64(data[i:])
-		tmp = t.decode(block ^ iv2)
-		iv2 = tmp
-		holder = tmp ^ iv1
-		iv1 = block
-		binary.BigEndian.PutUint64(dst[i:], holder)
+		holder = iv1
+		iv1 = binary.BigEndian.Uint64(data[i:])
+		iv2 = dectea(t0, t1, t2, t3, iv1^iv2)
+		binary.BigEndian.PutUint64(dst[i:], iv2^holder)
 	}
 	return dst[dst[0]&7+3 : len(data)-7]
 }
 
-var sumTable = [0x10]uint32{
-	0x9e3779b9,
-	0x3c6ef372,
-	0xdaa66d2b,
-	0x78dde6e4,
-	0x1715609d,
-	0xb54cda56,
-	0x5384540f,
-	0xf1bbcdc8,
-	0x8ff34781,
-	0x2e2ac13a,
-	0xcc623af3,
-	0x6a99b4ac,
-	0x08d12e65,
-	0xa708a81e,
-	0x454021d7,
-	0xe3779b90,
-}
-
 //go:nosplit
-func (t *TEA) encode(n uint64) uint64 {
+func enctea(t0, t1, t2, t3 uint32, n uint64) uint64 {
 	v0, v1 := uint32(n>>32), uint32(n)
-	for i := 0; i < 0x10; i++ {
-		v0 += ((v1 << 4) + t[0]) ^ (v1 + sumTable[i]) ^ ((v1 >> 5) + t[1])
-		v1 += ((v0 << 4) + t[2]) ^ (v0 + sumTable[i]) ^ ((v0 >> 5) + t[3])
-	}
+
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x9e3779b9) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x9e3779b9) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x3c6ef372) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x3c6ef372) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0xdaa66d2b) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0xdaa66d2b) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x78dde6e4) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x78dde6e4) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x1715609d) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x1715609d) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0xb54cda56) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0xb54cda56) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x5384540f) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x5384540f) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0xf1bbcdc8) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0xf1bbcdc8) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x8ff34781) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x8ff34781) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x2e2ac13a) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x2e2ac13a) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0xcc623af3) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0xcc623af3) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x6a99b4ac) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x6a99b4ac) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x08d12e65) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x08d12e65) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0xa708a81e) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0xa708a81e) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0x454021d7) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0x454021d7) ^ ((v0 >> 5) + t3)
+	v0 += ((v1 << 4) + t0) ^ (v1 + 0xe3779b90) ^ ((v1 >> 5) + t1)
+	v1 += ((v0 << 4) + t2) ^ (v0 + 0xe3779b90) ^ ((v0 >> 5) + t3)
+
 	return uint64(v0)<<32 | uint64(v1)
 }
 
 // 每次8字节
 //go:nosplit
-func (t *TEA) decode(n uint64) uint64 {
+func dectea(t0, t1, t2, t3 uint32, n uint64) uint64 {
 	v0, v1 := uint32(n>>32), uint32(n)
-	t0, t1, t2, t3 := t[0], t[1], t[2], t[3]
 
 	v1 -= ((v0 << 4) + t2) ^ (v0 + 0xe3779b90) ^ ((v0 >> 5) + t3)
 	v0 -= ((v1 << 4) + t0) ^ (v1 + 0xe3779b90) ^ ((v1 >> 5) + t1)
@@ -89,8 +99,8 @@ func (t *TEA) decode(n uint64) uint64 {
 	v0 -= ((v1 << 4) + t0) ^ (v1 + 0x454021d7) ^ ((v1 >> 5) + t1)
 	v1 -= ((v0 << 4) + t2) ^ (v0 + 0xa708a81e) ^ ((v0 >> 5) + t3)
 	v0 -= ((v1 << 4) + t0) ^ (v1 + 0xa708a81e) ^ ((v1 >> 5) + t1)
-	v1 -= ((v0 << 4) + t2) ^ (v0 + 0x8d12e65) ^ ((v0 >> 5) + t3)
-	v0 -= ((v1 << 4) + t0) ^ (v1 + 0x8d12e65) ^ ((v1 >> 5) + t1)
+	v1 -= ((v0 << 4) + t2) ^ (v0 + 0x08d12e65) ^ ((v0 >> 5) + t3)
+	v0 -= ((v1 << 4) + t0) ^ (v1 + 0x08d12e65) ^ ((v1 >> 5) + t1)
 	v1 -= ((v0 << 4) + t2) ^ (v0 + 0x6a99b4ac) ^ ((v0 >> 5) + t3)
 	v0 -= ((v1 << 4) + t0) ^ (v1 + 0x6a99b4ac) ^ ((v1 >> 5) + t1)
 	v1 -= ((v0 << 4) + t2) ^ (v0 + 0xcc623af3) ^ ((v0 >> 5) + t3)