```
name         old time/op    new time/op    delta
TEAen/16-8      273ns ± 4%     264ns ± 1%  -3.37%  (p=0.008 n=5+5)
TEAen/256-8    1.96µs ± 1%    1.87µs ± 0%  -4.56%  (p=0.008 n=5+5)
TEAen/4K-8     28.1µs ± 5%    27.5µs ± 0%    ~     (p=0.056 n=5+5)
TEAen/32K-8     222µs ± 1%     223µs ± 1%    ~     (p=0.095 n=5+5)
TEAde/16-8      220ns ± 1%     220ns ± 0%    ~     (p=0.690 n=5+5)
TEAde/256-8    1.75µs ± 0%    1.74µs ± 1%  -0.80%  (p=0.008 n=5+5)
TEAde/4K-8     26.0µs ± 0%    25.8µs ± 1%  -0.81%  (p=0.016 n=5+5)
TEAde/32K-8     211µs ± 0%     209µs ± 0%  -1.24%  (p=0.008 n=5+5)

name         old speed      new speed      delta
TEAen/16-8   58.7MB/s ± 4%  60.7MB/s ± 1%  +3.44%  (p=0.008 n=5+5)
TEAen/256-8   131MB/s ± 1%   137MB/s ± 0%  +4.78%  (p=0.008 n=5+5)
TEAen/4K-8    146MB/s ± 5%   149MB/s ± 0%    ~     (p=0.056 n=5+5)
TEAen/32K-8   148MB/s ± 1%   147MB/s ± 1%    ~     (p=0.095 n=5+5)
TEAde/16-8    145MB/s ± 1%   146MB/s ± 0%    ~     (p=0.690 n=5+5)
TEAde/256-8   155MB/s ± 0%   157MB/s ± 1%  +0.81%  (p=0.008 n=5+5)
TEAde/4K-8    158MB/s ± 0%   160MB/s ± 1%  +0.82%  (p=0.016 n=5+5)
TEAde/32K-8   155MB/s ± 0%   157MB/s ± 0%  +1.26%  (p=0.008 n=5+5)
```